+++ utils/create_voc_step1.cpp
@@ -1,4 +1,10 @@
-//Second step,creates the vocabulary from the set of features. It can be slow
+/**
+ * Date:  2016
+ * Author: Rafael Muñoz Salinas
+ * Description: demo application of DBoW3
+ * License: see the LICENSE.txt file
+ */
+
 #include <iostream>
 #include <vector>
 
@@ -24,7 +30,7 @@ vector<cv::Mat> features;
     uint32_t size;
     ifile.read((char*)&size,sizeof(size));
     features.resize(size);
-    for(size_t i=0;i<size;i++){
+    for(int i=0;i<size;i++){
 
         uint32_t cols,rows,type;
         ifile.read( (char*)&cols,sizeof(cols));
